iota_test_transaction_builder

Function publish_example_package

pub async fn publish_example_package(
    context: &WalletContext,
    example_subpath: &'static str,
    sender_key_pair: &AccountKeyPair,
    sender: IotaAddress,
    gas: ObjectRef,
) -> (ObjectID, TransactionDigest)

Executes a transaction to publish the specified examples package and returns the package id and the digest of the transaction.